+++ src/timer.rs
@@ -2,10 +2,10 @@
 use crate::hal::timer::{CountDown, Periodic};
 use crate::pac::{tim2, tim21, tim22, tim6, TIM2, TIM21, TIM22, TIM3, TIM6};
 use crate::rcc::{Clocks, Rcc};
-use crate::time::Hertz;
 use cast::{u16, u32};
 use cortex_m::peripheral::syst::SystClkSource;
 use cortex_m::peripheral::SYST;
+use embedded_time::rate::Hertz;
 use void::Void;
 
 pub trait TimerExt<TIM> {